Refactor map initialization and layer control
This commit refactors the code related to map initialization and layer control in the index.html file. The map is now initialized with OpenStreetMap as the base map, and a layer switcher control is added. This change allows for the possibility of adding more layers depending on the data to be shown.